+++ packages/core/CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - feat(#334): Deprecate `core.onChangeMocks` method. `core.mock.onChange` method must be used instead.
 - feat(#334): Deprecate `core.loadMocks` method. `core.mock.createLoaders` method must be used instead.
 - feat(#334): Deprecate `core.loadRoutes` method. `core.mock.createLoaders` method must be used instead.
+- feat(#334): Deprecate `core.mock.plainRoutes` getter. `core.mock.routes.plain` getter must be used instead.
 
 ## [3.5.0] - 2022-07-05